Add stroke around the regions.
To do so, I had to do a few changes to the SVG. As there's no z-index in SVG
(later elements appear in front of everything else), the final structure had to
be:

  Region's <path>
  District's <path>
  Region's stroke <path>

The region has fill and receives the mouseclick events. The districts are drawn
over it, so their stroke appear on top. I couldn't change their drawing order,
because then the district would be unreachable. So I added a copy of the region
on top of everything, with just the stroke (fill: none). It doesn't grab any
mouse events, and everything works fine.

It probably has performance issues, but seems OK for now.