Fixes a bug where the tooltips wheren't showing.
This happened with the districts with dataset because, when we moved them to
the end of the SVG, we lost their parent's <a> element. This is the elemnt that
Kartograph.js uses to hold the tooltip.

To fix, instead of moving just the <path>, we move the containing <a> element.