Yearly Change columns aren't sortable.
This is because the table is loaded dynamically, but the
yearlyChange/yearlyChangePct fields are calculated client-side. So, as the code
is now, it's not possible to make them sortable.

They must either be added to the backend, or the table should be sortable on
the frontend. Neither is easy to do :/